+++ src/ui/marker.test.ts
@@ -157,13 +157,15 @@ describe('marker', () => {
         expect(!marker.getPopup()).toBeTruthy();
     });
 
-    test('Marker#togglePopup opens a popup that was closed', () => {
+    test('Marker#togglePopup opens a popup that was closed', async () => {
         const map = createMap();
         const marker = new Marker()
             .setLngLat([0, 0])
             .addTo(map)
-            .setPopup(new Popup())
-            .togglePopup();
+            .setPopup(new Popup());
+
+        await sleep(500);
+        marker.togglePopup();
 
         expect(marker.getPopup().isOpen()).toBeTruthy();
 
@@ -184,12 +186,14 @@ describe('marker', () => {
         map.remove();
     });
 
-    test('Enter key on Marker opens a popup that was closed', () => {
+    test('Enter key on Marker opens a popup that was closed', async () => {
         const map = createMap();
         const marker = new Marker()
             .setLngLat([0, 0])
             .addTo(map)
             .setPopup(new Popup());
+
+        await sleep(500);
 
         // popup not initially open
         expect(marker.getPopup().isOpen()).toBeFalsy();
@@ -202,12 +206,14 @@ describe('marker', () => {
         map.remove();
     });
 
-    test('Space key on Marker opens a popup that was closed', () => {
+    test('Space key on Marker opens a popup that was closed', async () => {
         const map = createMap();
         const marker = new Marker()
             .setLngLat([0, 0])
             .addTo(map)
             .setPopup(new Popup());
+
+        await sleep(500);
 
         // popup not initially open
         expect(marker.getPopup().isOpen()).toBeFalsy();
@@ -292,13 +298,15 @@ describe('marker', () => {
         map.remove();
     });
 
-    test('Popup anchors around default Marker', () => {
+    test('Popup anchors around default Marker', async () => {
         const map = createMap();
 
         const marker = new Marker()
             .setLngLat([0, 0])
             .setPopup(new Popup().setText('Test'))
             .addTo(map);
+
+        await sleep(500);
 
         // open the popup
         marker.togglePopup();
@@ -361,13 +369,15 @@ describe('marker', () => {
         map.remove();
     });
 
-    test('Popup is opened at its marker position after marker is moved to another globe', () => {
+    test('Popup is opened at its marker position after marker is moved to another globe', async () => {
         const map = createMap({width: 3000});
 
         const marker = new Marker()
             .setLngLat([0, 0])
             .setPopup(new Popup().setText('Test'))
             .addTo(map);
+
+        await sleep(500);
 
         marker._pos = new Point(2999, 242);
         marker._lngLat = map.unproject(marker._pos);
@@ -377,7 +387,7 @@ describe('marker', () => {
         map.remove();
     });
 
-    test('Popup is re-opened at its marker position after marker is moved to another globe', () => {
+    test('Popup is re-opened at its marker position after marker is moved to another globe', async () => {
         const map = createMap({width: 3000});
 
         const marker = new Marker()
@@ -387,6 +397,7 @@ describe('marker', () => {
             .togglePopup()
             .togglePopup();
 
+        await sleep(500);
         marker._pos = new Point(2999, 242);
         marker._lngLat = map.unproject(marker._pos);
         marker.togglePopup();
@@ -1026,6 +1037,56 @@ describe('marker', () => {
         marker.setOpacity(undefined, '0.35');
 
         expect(marker.getElement().style.opacity).toMatch('0.35');
+        map.remove();
+    });
+
+    test('Removes an open popup when going behind 3d terrain', async () => {
+        const map = createMap();
+        const marker = new Marker()
+            .setLngLat([0, 0])
+            .addTo(map)
+            .setPopup(new Popup());
+
+        await sleep(500);
+        marker.togglePopup();
+
+        expect(marker._popup.isOpen()).toBeTruthy();
+
+        map.transform.lngLatToCameraDepth = () => .95; // Mocking distance to marker
+
+        map.terrain = {
+            getElevationForLngLatZoom: () => 0,
+            depthAtPoint: () => .92
+        } as any as Terrain;
+        map.fire('terrain');
+
+        await sleep(500);
+
+        expect(marker._popup?.isOpen()).toBeFalsy();
+        map.remove();
+    });
+
+    test('Does not open a popup when behind 3d terrain', async () => {
+        const map = createMap();
+        const marker = new Marker()
+            .setLngLat([0, 0])
+            .addTo(map)
+            .setPopup(new Popup());
+
+        map.transform.lngLatToCameraDepth = () => .95;
+
+        map.terrain = {
+            getElevationForLngLatZoom: () => 0,
+            depthAtPoint: () => .92
+        } as any as Terrain;
+        map.fire('terrain');
+
+        await sleep(500);
+
+        marker.togglePopup();
+
+        expect(marker._popup.isOpen()).toBeFalsy();
+
         map.remove();
     });