+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 ## main
 
 ### ✨ Features and improvements
+- Hide Popup when its parent Marker is behind terrain ([#3865](https://github.com/maplibre/maplibre-gl-js/pull/3865))
 - _...Add new stuff here..._
 
 ### 🐞 Bug fixes
+- Fix type definition on `localIdeographFontFamily` ([#3896](https://github.com/maplibre/maplibre-gl-js/pull/3896))
 - Fix unwanted panning changes at the end of a panning motion ([#3872](https://github.com/maplibre/maplibre-gl-js/issues/3872))
 - _...Add new stuff here..._